move `g3w-editor.js` into `index.js`
